requirements.txt
# Core dependencies
ollama

# Image processing
Pillow
opencv-python

# PDF processing
PyMuPDF
pdf2image

# Data handling
numpy
pandas

# Network and utilities
requests


# Optional: For better fonts in visualization
# pillow-fonts>=1.0.0

# Development dependencies (optional)
# pytest>=7.4.0
# black>=23.0.0
# flake8>=6.0.0